move close_task() to TaskRepository and get_user_leaderboard to UserRepository()
