+++ backend/oasst_backend/user_repository.py
@@ -1,8 +1,9 @@
 from typing import Optional
 
-from oasst_backend.models import ApiClient, User
+from oasst_backend.models import ApiClient, Message, User
 from oasst_shared.schemas import protocol as protocol_schema
-from sqlmodel import Session
+from oasst_shared.schemas.protocol import LeaderboardStats
+from sqlmodel import Session, func
 
 
 class UserRepository:
@@ -40,3 +41,24 @@ class UserRepository:
             self.db.add(user)
             self.db.commit()
         return user
+
+    def get_user_leaderboard(self, role: str) -> LeaderboardStats:
+        """
+        Get leaderboard stats for Messages created,
+        separate leaderboard for prompts & assistants
+
+        """
+        query = (
+            self.db.query(Message.user_id, User.username, User.display_name, func.count(Message.user_id))
+            .join(User, User.id == Message.user_id, isouter=True)
+            .filter(Message.deleted is not True, Message.role == role)
+            .group_by(Message.user_id, User.username, User.display_name)
+            .order_by(func.count(Message.user_id).desc())
+        )
+
+        result = [
+            {"ranking": i, "user_id": j[0], "username": j[1], "display_name": j[2], "score": j[3]}
+            for i, j in enumerate(query.all(), start=1)
+        ]
+
+        return LeaderboardStats(leaderboard=result)
